# packages/mtmai/mtmai-0.3.693.tar.gz/mtmai-0.3.693/mtmai/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mtmai",
    "description": "mtmai description(group)",
    "version": "0.3.693"
  },
  "paths": {
...
    "/api/v1/agent/agent_bootstrap": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Agent Bootstrap",
        "description": "\u83b7\u53d6 agent \u7684\u914d\u7f6e\uff0c\u7528\u4e8e\u524d\u7aef\u52a0\u8f7dagent\u7684\u914d\u7f6e",
        "operationId": "agent-agent_bootstrap",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentBootstrap"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/agent": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "\u83b7\u53d6 Agent \u5217\u8868",
        "description": "\u6b64\u7aef\u70b9\u7528\u4e8e\u83b7\u53d6 agent \u5217\u8868\u3002\u652f\u6301\u5206\u9875\u529f\u80fd\u53ef\u4ee5\u901a\u8fc7 `skip` \u548c `limit` \u53c2\u6570\u63a7\u5236\u8fd4\u56de\u7684 agent \u6570\u91cf\u3002",
        "operationId": "agent-items",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "\u6210\u529f\u8fd4\u56de agent \u5217\u8868",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentsPublic"
                },
                "example": {
                  "data": [
                    {
                      "name": "agent1",
                      "status": "active"
                    },
                    {
                      "name": "agent2",
                      "status": "inactive"
                    }
                  ],
                  "count": 2
                }
              }
            }
          },
          "401": {
            "description": "\u672a\u7ecf\u6388\u6743\u7684\u8bf7\u6c42"
          },
          "500": {
            "description": "\u670d\u52a1\u5668\u5185\u90e8\u9519\u8bef"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/{agent_id}": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Item",
        "operationId": "agent-get_item",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AgentMeta"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Agent-Get Item"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/image/{agent}": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "\u83b7\u53d6\u5de5\u4f5c\u6d41\u56fe\u50cf",
        "description": "\u6b64\u7aef\u70b9\u901a\u8fc7\u7ed9\u5b9a\u7684 agent ID\uff0c\u751f\u6210\u5de5\u4f5c\u6d41\u7684\u56fe\u50cf\u5e76\u8fd4\u56de PNG \u683c\u5f0f\u7684\u6570\u636e\u3002",
        "operationId": "agent-image",
        "responses": {
          "200": {
            "description": "\u8fd4\u56de PNG \u683c\u5f0f\u7684\u5de5\u4f5c\u6d41\u56fe\u50cf\u3002",
            "content": {
              "application/json": {
                "schema": {}
              },
              "image/png": {}
            }
          },
          "404": {
            "description": "Agent \u672a\u627e\u5230"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/state": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "\u83b7\u53d6\u5de5\u4f5c\u6d41\u72b6\u6001",
        "operationId": "agent-state",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "\u8fd4\u56de\u5de5\u4f5c\u6d41\u5f53\u524d\u5b8c\u6574\u72b6\u6001\u6570\u636e",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/completions": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Completions",
        "operationId": "agent-completions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/uistate": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Uistate",
        "description": "\u76ee\u524d\u4ec5\u4f5c\u5360\u4f4d\u7b26, \u8ba9\u524d\u7aef\u53ef\u4ee5\u901a\u8fc7 openapi \u8ddf\u83b7\u53d6\u5230ChatBotUiState \u7c7b\u578b",
        "operationId": "agent-uistate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatBotUiState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/run_agent_task": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Run Agent Task",
        "description": "agent \u4efb\u52a1\u8c03\u7528\u8bf7\u6c42",
        "operationId": "agent-run_agent_task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/mtmai__api__routes__agent__AgentTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/task_form": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Task Form",
        "description": "\u6839\u636e\u4efb\u52a1\u8bf7\u6c42\uff0c\u8fd4\u56de\u4efb\u52a1\u8868\u5355",
        "operationId": "agent-task_form",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskFormRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskFormResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentBootstrap": {
        "properties": {
          "view_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentViewType"
              },
              {
                "type": "null"
              }
            ],
            "default": "sidebar"
          },
          "is_show_fab": {
            "type": "boolean",
            "title": "Is Show Fab",
            "default": true
          }
        },
        "type": "object",
        "title": "AgentBootstrap",
        "description": "\u524d\u7aef\u83b7\u53d6 agent \u7684\u914d\u7f6e\n\u524d\u7aef\u5168\u5c40agent \u52a0\u8f7d\u5668\u4f1a\u5728\u6240\u6709\u9875\u9762\u52a0\u8f7d\u65f6\uff0c\u52a0\u8f7d\u4e00\u6b21\uff0c\u6839\u636e\u8fd4\u56de\u7684\u914d\u7f6e\uff0c\u521d\u59cb\u5316agent"
      },
      "AgentMeta": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "chat_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Url"
          },
          "can_chat": {
            "type": "boolean",
            "title": "Can Chat",
            "default": [
              false
            ]
          },
          "agent_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Type"
          },
          "graph_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph Image"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "base_url"
        ],
        "title": "AgentMeta"
      },
      "AgentStateRequest": {
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          }
        },
        "type": "object",
        "required": [
          "thread_id"
        ],
        "title": "AgentStateRequest"
      },
...
      "AgentTaskRequest": {
        "properties": {
          "task_name": {
            "type": "string",
            "title": "Task Name"
          }
        },
        "type": "object",
        "required": [
          "task_name"
        ],
        "title": "AgentTaskRequest"
      },
      "AgentViewType": {
        "type": "string",
        "enum": [
          "sidebar",
          "popup"
        ],
        "title": "AgentViewType"
      },
      "AgentsPublic": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AgentMeta"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "AgentsPublic"
      },
...
      "ChatBotUiState": {
        "properties": {
          "agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent"
          },
          "layout": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layout"
          },
          "theme": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme"
          },
          "threadId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threadid"
          },
          "config": {
            "type": "object",
            "title": "Config"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "title": "ChatBotUiState"
      },
      "CommonFormData": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CommonFormField"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          }
        },
        "type": "object",
        "title": "CommonFormData"
      },
      "CommonFormField": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required"
          },
          "default": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "placeholder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placeholder"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CommonFormField"
      },
      "CompletinRequest": {
        "properties": {
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "option": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Option"
          },
          "task": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "CompletinRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskFormRequest": {
        "properties": {
          "task_name": {
            "type": "string",
            "title": "Task Name"
          }
        },
        "type": "object",
        "required": [
          "task_name"
        ],
        "title": "TaskFormRequest"
      },
      "TaskFormResponse": {
        "properties": {
          "form": {
            "$ref": "#/components/schemas/CommonFormData"
          }
        },
        "type": "object",
        "required": [
          "form"
        ],
        "title": "TaskFormResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "mtmai__api__routes__agent__AgentTaskResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "result": {
            "type": "string",
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "result"
        ],
        "title": "AgentTaskResponse"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  },
...
}